backend/requirements.txt
fastapi>=0.108.0
uvicorn[standard]>=0.25.0
python-multipart>=0.0.6
PyPDF2>=3.0.1
python-dotenv>=1.0.0
google-generativeai>=0.3.2
pydantic>=2.7.4
pydantic-settings>=2.1.0
httpx>=0.26.0
aiofiles>=23.2.1
pypdf>=3.17.4
requests>=2.31.0

# LangGraph and LangChain for multi-agent workflows
langgraph>=0.2.28
langchain>=0.3.7
langchain-google-genai>=2.0.0
langchain-community>=0.3.5

# Tools for verification
beautifulsoup4>=4.12.3
duckduckgo-search>=4.4.3
wikipedia>=1.4.0
